platzi-ecommerce-backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==45.0.3
dataclasses-json==0.6.7
Faker==37.3.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
firebase-admin==6.9.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
fsspec==2025.3.2
gcloud==0.18.3
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio==5.25.2
gradio_client==1.8.0
groovy==0.1.2
grpcio==1.72.1
grpcio-status==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jwcrypto==1.5.6
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.52
langchain-text-splitters==0.3.8
langsmith==0.3.31
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
msgpack==1.1.0
multidict==6.4.3
mypy-extensions==1.0.0
numpy==2.2.4
oauth2client==4.1.3
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jwt==4.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==0.10.1
rich==14.0.0
rsa==4.9.1
ruff==0.11.5
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tavily-python==0.5.4
tenacity==9.1.2
tiktoken==0.9.0
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.20
uvicorn==0.34.1
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.0
zstandard==0.23.0
